Extrait une fonction de test
Permet d'importer `getAuthUserFromUserCredentials` dans une fixture utilisables aussi dans les tests e2e ui.
Ce n'était pas possible de le faire depuis `@tet/backend/test` car `test/index.ts` réexporte `app-utils`, qui charge toute l’app Nest, mais Playwright transpile sans support des décorateurs de paramètres ce qui conduit à une erreur.
